requirements.txt
appnope==0.1.0
asn1crypto==0.24.0
backcall==0.1.0
bcrypt==3.1.6
boto3==1.9.120
botocore==1.12.120
cffi==1.12.2
cryptography==2.6.1
decorator==4.4.0
docutils==0.14
ipython==7.4.0
ipython-genutils==0.2.0
jedi==0.13.3
jmespath==0.9.4
kafka-python==1.4.5
paramiko==2.4.2
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pyasn1==0.4.5
pycparser==2.19
Pygments==2.3.1
PyNaCl==1.3.0
python-dateutil==2.8.0
s3transfer==0.2.0
six==1.12.0
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
